Separated synchronization utilities from filesystem
