+++ client/src/shared/utils/synchronization/exportNotebookStructure.ts
@@ -1,4 +1,4 @@
-import { saveFile } from './saveFile';
+import { saveFile } from '../fileSystem/saveFile';
 import { FileItem } from '../../types';
 import { getNotebookStructure } from './getNotebookStructure';
