+++ client/src/shared/utils/synchronization/getNotebookStructure.ts
